修复bug
1、兼容64位
2、修复了输入框避免遮挡移动视图错误的bug
3、修复了在iOS7.1.1下显示状态信息框崩溃的bug

+++ pjcore/utilities/DESUtility.m
@@ -80,7 +80,7 @@
 +(NSString *)encryptWithMD5:(NSString *)string{
     const char    *characters=[string UTF8String];
     unsigned char result[CC_MD5_DIGEST_LENGTH];
-    int charCount=strlen(characters);
+    int charCount=(int)strlen(characters);
     
     CC_MD5(characters, charCount, result);
